Commit made by the Bioconductor Git-SVN bridge. Consists of 3 commits.
Commit information:

Commit id: ce26f1d519ca5e35d47fd1d8cfd5eae6fb8a5a10
Commit message: fix missed size_t typo
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 16:02:25 -0700
Author date: 2014-08-31 16:02:25 -0700

Commit id: 1bf9bb8a913d0bd7a2ae760afc84c5c223591602
Commit message: change the linking interfaces to match changes to code
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 15:58:09 -0700
Author date: 2014-08-31 15:58:09 -0700

Commit id: 832752c8a9b4dc0c2eb846db7240487fd36b6ceb
Commit message: change, where appropriate, to size_t
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 15:50:19 -0700
Author date: 2014-08-31 15:50:19 -0700



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/preprocessCore@93723 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/avg_log.h
@@ -1,10 +1,10 @@
 #ifndef AVG_LOG_H
 #define AVG_LOG_H
 
-void AverageLog(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes, double *resultsSE);
-void AverageLog_noSE(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes);
+void AverageLog(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes, double *resultsSE);
+void AverageLog_noSE(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes);
 
-void averagelog_no_copy(double *data, int rows, int cols, double *results, double *resultsSE);
-void averagelog(double *data, int rows, int cols, double *results, double *resultsSE);
+void averagelog_no_copy(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
+void averagelog(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
 
 #endif
